airflow/requirements.txt
apache-airflow
pyspark
pandas
pandas-gbq
requests
google-cloud-storage
google-cloud-bigquery
gspread
gcsfs